Remove code duplication between numpy and autograd backends by adding _shared_numpy folder and dispatch at import time
