+++ geomstats/_backend/_shared_numpy/_dispatch.py
@@ -1,0 +1,13 @@
+from geomstats._backend import BACKEND_NAME
+
+if BACKEND_NAME == "numpy":
+    import numpy
+    import scipy
+
+    from ..numpy import _common
+elif BACKEND_NAME == "autograd":
+    from autograd import numpy, scipy
+
+    from ..autograd import _common
+else:
+    raise ValueError("Unknown numpy.")
