+++ geomstats/_backend/autograd/linalg.py
@@ -3,10 +3,9 @@
 import functools as _functools
 
 import autograd.numpy as _np
-import autograd.scipy as _ascp
 from autograd.extend import defvjp as _defvjp
 from autograd.extend import primitive as _primitive
-from autograd.numpy.linalg import (  # NOQA
+from autograd.numpy.linalg import (
     cholesky,
     det,
     eig,
@@ -20,20 +19,9 @@ from autograd.numpy.linalg import (  # NOQA
 )
 from autograd.scipy.linalg import expm
 
-from ._common import atol
-from ._common import to_ndarray as _to_ndarray
-from ._dtype import _cast_fout_to_input_dtype, _cast_out_to_input_dtype
-
-_diag_vec = _np.vectorize(_np.diag, signature="(n)->(n,n)")
-
-_logm_vec = _cast_fout_to_input_dtype(
-    target=_np.vectorize(_ascp.linalg.logm, signature="(n,m)->(n,m)")
-)
-
-
-def _is_symmetric(x, tol=1e-12):
-    new_x = _to_ndarray(x, to_ndim=3)
-    return (_np.abs(new_x - _np.transpose(new_x, axes=(0, 2, 1))) < tol).all()
+from .._shared_numpy.linalg import fractional_matrix_power, is_single_matrix_pd
+from .._shared_numpy.linalg import logm as _logm
+from .._shared_numpy.linalg import qr, quadratic_assignment, solve_sylvester, sqrtm
 
 
 def _adjoint(_ans, x, fn):
@@ -56,84 +44,7 @@ _expm_vjp = _functools.partial(_adjoint, fn=expm)
 _defvjp(expm, _expm_vjp)
 
 
-@_primitive
-def logm(x):
-    ndim = x.ndim
-    new_x = _to_ndarray(x, to_ndim=3)
-    if _is_symmetric(new_x):
-        eigvals, eigvecs = _np.linalg.eigh(new_x)
-        if (eigvals > 0).all():
-            eigvals = _np.log(eigvals)
-            eigvals = _diag_vec(eigvals)
-            transp_eigvecs = _np.transpose(eigvecs, axes=(0, 2, 1))
-            result = _np.matmul(eigvecs, eigvals)
-            result = _np.matmul(result, transp_eigvecs)
-        else:
-            result = _logm_vec(new_x)
-    else:
-        result = _logm_vec(new_x)
-
-    if ndim == 2:
-        return result[0]
-    return result
-
+logm = _primitive(_logm)
 
 _logm_vjp = _functools.partial(_adjoint, fn=logm)
 _defvjp(logm, _logm_vjp)
-
-
-def solve_sylvester(a, b, q):
-    if a.shape == b.shape:
-        axes = (0, 2, 1) if a.ndim == 3 else (1, 0)
-        if _np.all(a == b) and _np.all(_np.abs(a - _np.transpose(a, axes)) < 1e-12):
-            eigvals, eigvecs = eigh(a)
-            if _np.all(eigvals >= 1e-12):
-                tilde_q = _np.transpose(eigvecs, axes) @ q @ eigvecs
-                tilde_x = tilde_q / (eigvals[..., :, None] + eigvals[..., None, :])
-                return eigvecs @ tilde_x @ _np.transpose(eigvecs, axes)
-
-    return _np.vectorize(
-        _ascp.linalg.solve_sylvester, signature="(m,m),(n,n),(m,n)->(m,n)"
-    )(a, b, q)
-
-
-@_cast_fout_to_input_dtype
-def sqrtm(x):
-    return _np.vectorize(_ascp.linalg.sqrtm, signature="(n,m)->(n,m)")(x)
-
-
-def quadratic_assignment(a, b, options):
-    return list(_ascp.optimize.quadratic_assignment(a, b, options=options).col_ind)
-
-
-def qr(*args, **kwargs):
-    return _np.vectorize(
-        _np.linalg.qr, signature="(n,m)->(n,k),(k,m)", excluded=["mode"]
-    )(*args, **kwargs)
-
-
-def is_single_matrix_pd(mat):
-    """Check if a 2D square matrix is positive definite."""
-    if mat.shape[0] != mat.shape[1]:
-        return False
-    if mat.dtype in [_np.complex64, _np.complex128]:
-        is_hermitian = _np.all(_np.abs(mat - _np.conj(_np.transpose(mat))) < atol)
-        if not is_hermitian:
-            return False
-        eigvals = _np.linalg.eigvalsh(mat)
-        return _np.min(_np.real(eigvals)) > 0
-    try:
-        _np.linalg.cholesky(mat)
-        return True
-    except _np.linalg.LinAlgError as e:
-        if e.args[0] == "Matrix is not positive definite":
-            return False
-        raise e
-
-
-@_cast_out_to_input_dtype
-def fractional_matrix_power(A, t):
-    if A.ndim == 2:
-        return _ascp.linalg.fractional_matrix_power(A, t)
-
-    return _np.stack([_ascp.linalg.fractional_matrix_power(A_, t) for A_ in A])
