+++ geomstats/_backend/autograd/random.py
@@ -1,29 +1,5 @@
 """Autograd based random backend."""
 import autograd.numpy as _np
-from autograd.numpy.random import default_rng as _default_rng  # NOQA
 from autograd.numpy.random import randint, seed
 
-from ._dtype import _allow_complex_dtype, _modify_func_default_dtype
-
-rand = _modify_func_default_dtype(
-    copy=False, kw_only=True, target=_allow_complex_dtype(target=_np.random.rand)
-)
-
-uniform = _modify_func_default_dtype(
-    copy=False, kw_only=True, target=_allow_complex_dtype(target=_np.random.uniform)
-)
-
-
-normal = _modify_func_default_dtype(
-    copy=False, kw_only=True, target=_allow_complex_dtype(target=_np.random.normal)
-)
-
-multivariate_normal = _modify_func_default_dtype(
-    copy=False,
-    kw_only=True,
-    target=_allow_complex_dtype(target=_np.random.multivariate_normal),
-)
-
-
-def choice(*args, **kwargs):
-    return _default_rng().choice(*args, **kwargs)
+from .._shared_numpy.random import choice, multivariate_normal, normal, rand, uniform
